Switch to boto3 for SQS
Also work in batch of 10 messages (limit from Amazon) in order to go
faster. We currently loose a lot of time just sending and receiving
messages from SQS.

+++ tilecloud/store/sqs.py
@@ -1,11 +1,12 @@
+import base64
+import json
 import logging
 import time
-
-from boto.exception import SQSDecodeError, SQSError
 
 from tilecloud import Tile, TileCoord, TileStore
 
 
+BATCH_SIZE = 10  # max Amazon allows
 logger = logging.getLogger(__name__)
 
 
@@ -34,25 +35,27 @@ class SQSTileStore(TileStore):
 
     def list(self):
         while True:
-            try:
-                sqs_message = self.queue.read()
-                if sqs_message is None:
+            sqs_messages = self.queue.receive_messages(MaxNumberOfMessages=BATCH_SIZE)
+            if not sqs_messages:
+                try:
+                    self.on_empty(self.queue)
+                except StopIteration:
+                    break
+            else:
+                for sqs_message in sqs_messages:
                     try:
-                        self.on_empty(self.queue)
-                    except StopIteration:
-                        break
-                else:
-                    z = sqs_message.get('z')
-                    x = sqs_message.get('x')
-                    y = sqs_message.get('y')
-                    n = sqs_message.get('n')
-                    metadata = sqs_message.get('metadata', {})
-                    # FIXME deserialize other attributes
-                    tile = Tile(TileCoord(z, x, y, n), sqs_message=sqs_message, metadata=metadata)
-                    yield tile
-            except SQSDecodeError as e:
-                logger.warning(str(e))
-                sqs_message.delete()
+                        body = json.loads(base64.b64decode(sqs_message.body.encode('utf-8')).decode('utf-8'))
+                        z = body.get('z')
+                        x = body.get('x')
+                        y = body.get('y')
+                        n = body.get('n')
+                        metadata = body.get('metadata', {})
+                        # FIXME deserialize other attributes
+                        tile = Tile(TileCoord(z, x, y, n), sqs_message=sqs_message, metadata=metadata)
+                        yield tile
+                    except Exception:
+                        logger.warning('Failed decoding the SQS message', exc_info=True)
+                        sqs_message.delete()
 
     @staticmethod
     def delete_one(tile):
@@ -62,19 +65,52 @@ class SQSTileStore(TileStore):
         return tile
 
     def put_one(self, tile):
-        sqs_message = self.queue.new_message()
-        sqs_message['z'] = tile.tilecoord.z
-        sqs_message['x'] = tile.tilecoord.x
-        sqs_message['y'] = tile.tilecoord.y
-        sqs_message['n'] = tile.tilecoord.n
-        sqs_message['metadata'] = tile.metadata
-        if 'sqs_message' in sqs_message['metadata']:
-            del sqs_message['metadata']['sqs_message']
+        sqs_message = _create_message(tile)
 
-        # FIXME serialize other attributes
         try:
-            self.queue.write(sqs_message)
-            tile.sqs_message = sqs_message
-        except SQSError as e:
+            self.queue.send_message(MessageBody=sqs_message)
+        except Exception as e:
+            logger.warning('Failed sending SQS message', exc_info=True)
             tile.error = e
         return tile
+
+    def put(self, tiles):
+        buffered_tiles = []
+        for tile in tiles:
+            buffered_tiles.append(tile)
+            if len(buffered_tiles) >= BATCH_SIZE:
+                self._send_buffer(buffered_tiles)
+                buffered_tiles = []
+        if len(buffered_tiles) > 0:
+            self._send_buffer(buffered_tiles)
+
+    def _send_buffer(self, tiles):
+        try:
+            messages = [{
+                'Id': str(i),
+                'MessageBody': _create_message(tile)
+            } for i, tile in enumerate(tiles)]
+            response = self.queue.send_messages(Entries=messages)
+            for failed in response.get('Failed', []):
+                logger.warning('Failed sending SQS message: %s', failed['Message'])
+                pos = int(failed['Id'])
+                tiles[pos].error = failed['Message']
+        except Exception as e:
+            logger.warning('Failed sending SQS messages', exc_info=True)
+            for tile in tiles:
+                tile.error = e
+
+
+def _create_message(tile):
+    sqs_message = {
+        'z': tile.tilecoord.z,
+        'x': tile.tilecoord.x,
+        'y': tile.tilecoord.y,
+        'n': tile.tilecoord.n,
+        'metadata': tile.metadata
+    }
+    if 'sqs_message' in sqs_message['metadata']:
+        del sqs_message['metadata']['sqs_message']
+
+    # FIXME serialize other attributes
+    return base64.b64encode(json.dumps(sqs_message).encode('utf-8')).decode('utf-8')
